refactor: move FormalFifoTester
